installer: improve region detection reliability using multiple APIs
The installation script has been updated to check multiple geolocation APIs
in order to determine the user's region more reliably. It now attempts to
query a list of URLs, including Cloudflare's DNS service, in case the primary
API is unreachable or returns incomplete data. This change addresses issues
with previous single-source IP detection and helps to ensure that the correct
download link is offered to the user based on their location.